Update F. Digits Summation.cpp

+++ C++ Assignments/Codeforces/Sheet1/F. Digits Summation.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main()
 {
-  int N,M;
+ long long int N,M;
   cin>>N>>M;
   cout<<(N%10)+(M%10);
